pyproject.toml
[project]
name = "wiki-langgraph"
version = "0.1.0"
description = "LangGraph pipeline: ingest sources, compile Obsidian-style wiki, index, CLI agent hooks"
readme = "README.md"
authors = [{ name = "wiki-langgraph contributors" }]
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0",
]

[project.scripts]
wiki-langgraph = "wiki_langgraph.cli:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
